+++ sync-for-payables/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [typescript v1.1.0] sync-for-payables
 ### Releases
 - [NPM v1.1.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/1.1.0 - sync-for-payables
+
+## 2023-09-22 15:33:32
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.90.0 (2.125.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.2.0] sync-for-payables
+### Releases
+- [NPM v1.2.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/1.2.0 - sync-for-payables